+++ CHANGELOG.md
@@ -42,6 +42,7 @@ Introduces save state file version 14.
 ### Fixes
 
 - App: Don't persist window geometry in full screen mode.
+- App: Reset window geometry if it matches the bounds of a display.
 - Backup RAM: Fix crash when attempting to load a backup RAM cartridge with the default path.
 - CD Block (HLE): Report current CD status and raise CMOK HIRQ signal when processing unimplemented commands.
 - CD Block (HLE): Read reset position flag correctly from parameters.